requirements.txt
numpy
python >= 3.7
pytorch == 1.9
simplejson
tensorboard
yacs
timm